NRFX-NONE: Fix compilation warning about intergers conversion
In nrf_cache.h there is an arithmetic opearation that causes
a warning during build. It is related with integer coversions.
Compiler wanrs about possible lost of data.

The operation calculates index of a data word in cache line.
The ouptus of the operation can't be larger than NRF_CACHEDATA-
_DATA_WORDS_IN_UNIT_MAX that is unsinged int but its actual value
fits into uint8_t. What more the word argument is checked before
against NRF_CACHEDATA_WORD_INDEX_MAX if is it smaller.
That together makes sure the end value falls into right range.

+++ hal/nrf_cache.h
@@ -656,7 +656,7 @@ NRF_STATIC_INLINE uint32_t nrf_cache_data_get(NRF_CACHEDATA_Type const * p_reg,
     NRFX_ASSERT(set < NRF_CACHEDATA_SET_INDEX_MAX);
 
     uint8_t du   = (word / NRF_CACHEDATA_DATA_WORDS_IN_UNIT_MAX);
-    uint8_t data = (word - (du * NRF_CACHEDATA_DATA_WORDS_IN_UNIT_MAX));
+    uint8_t data = (uint8_t)(word - (du * NRF_CACHEDATA_DATA_WORDS_IN_UNIT_MAX));
 
     return p_reg->SET[set].WAY[way].DU[du].DATA[data];
 #else
